moved server part to docker, replaced node.js static handling to nginx, migration of backend to 7th babel
