+++ addons/debug/Panel.jsx
