+++ addons/debug/package.json
