+++ addons/debug/webpack.config.js
