+++ addons/inventory/index.html
