+++ addons/inventory/main.js
