+++ addons/inventory/manifest.json
