+++ addons/inventory/package.json
