+++ addons/main-panel/index.html
