+++ addons/main-panel/main.js
