+++ addons/main-panel/manifest.json
