+++ addons/main-panel/package.json
