+++ addons/sw-chat/main.js
