+++ addons/sw-chat/manifest.json
