+++ addons/sw-chat/package.json
