+++ addons/sw-minimap/index.html
