+++ addons/sw-minimap/main.js
