+++ addons/sw-minimap/manifest.json
