+++ addons/sw-minimap/package.json
