+++ addons/unit-player/manifest.json
