+++ addons/xpbar/index.js
