+++ addons/xpbar/manifest.json
