+++ addons/xpbar/package.json
