+++ addons/xpbar/scripts/directives.js
