+++ docker/nginx/default.conf
@@ -1,0 +1,8 @@
+server {
+  listen 80;
+  server_name localhost;
+  location / {
+    root   /usr/share/nginx/html;
+    try_files $uri $uri/ $uri.html =404;
+  }
+}
