+++ package.json
@@ -4,8 +4,6 @@
   "private": true,
   "dependencies": {
     "babel-eslint": "^8.2.6",
-    "express": "^4.16.3",
-    "fs-extra": "^7.0.0",
     "gl-matrix": "^2.7.1",
     "pako": "^1.0.6",
     "react": "^16.4.2",
@@ -16,19 +14,20 @@
     "redux-devtools-extension": "^2.13.5",
     "redux-thunk": "^2.3.0",
     "seedrandom": "^2.4.4",
-    "simplex-noise": "^2.4.0",
     "uuid": "^3.3.2",
     "ws": "^6.0.0",
     "zlib": "^1.0.5"
   },
   "scripts": {
     "start": "react-app-rewired start",
-    "start-server": "babel-node app.js",
+    "start-server": "docker-compose up",
+    "rebuild-server": "docker-compose up -d --force-recreate --build",
     "build": "react-app-rewired build",
     "test": "react-app-rewired test --env=jsdom",
     "eject": "react-app-rewired eject",
     "flow": "./node_modules/flow-bin/cli.js",
-    "lint": "./node_modules/eslint/bin/eslint.js --ignore-pattern=node_modules --ignore-pattern=bundle.js src server"
+    "lint": "./node_modules/eslint/bin/eslint.js --ignore-pattern=node_modules --ignore-pattern=bundle.js src server",
+    "fix-nvm-sudo": "n=$(which node);n=${n%/bin/node}; chmod -R 755 $n/bin/*; sudo cp -r $n/{bin,lib,share} /usr/local"
   },
   "proxy": "http://localhost:1337",
   "devDependencies": {
