+++ server/index.js
@@ -1,3 +1,4 @@
+// @flow
 import Server from './server';
 
 export default new Server();
